requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.17.2
APScheduler==3.11.2
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
blinker==1.9.0
Brotli==1.1.0
cached-property==2.0.1
cachelib==0.13.0
cairocffi==1.7.1
certifi==2025.4.26
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==3.4.8
cssselect2==0.8.0
decorator==5.2.1
dnspython==2.7.0
docopt==0.6.2
docutils==0.21.2
email_validator==2.2.0
et_xmlfile==2.0.0
Flask==3.1.1
Flask-APScheduler==1.13.1
Flask-Caching==2.3.1
flask-cors==6.0.2
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.58.1
fpdf==1.7.2
frozenlist==1.7.0
geographiclib==2.0
geopy==2.4.1
gevent==25.9.1
greenlet==3.2.2
gunicorn==20.1.0
hiredis==3.3.0
idna==3.10
iniconfig==2.3.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
libsass==0.23.0
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.2
maxminddb==2.7.0
msgspec==0.20.0
multidict==6.6.3
num2words==0.5.10
numpy==2.2.6
ofxparse==0.21
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pdfkit==1.0.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
polib==1.1.1
propcache==0.3.2
psutil==7.2.1
psycopg2==2.9.11
psycopg2-binary==2.9.11
pycparser==2.22
pydot==1.4.2
pydyf==0.11.0
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.1
pyOpenSSL==21.0.0
pyparsing==3.2.3
PyPDF2==1.26.0
pyphen==0.17.2
pypiwin32==223
pyserial==3.5
pytest==9.0.2
python-barcode==0.16.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-stdnum==1.17
pytz==2025.2
pyusb==1.2.1
pywin32==310
qrcode==7.3.1
redis==7.1.0
reportlab==4.4.7
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rjsmin==1.1.0
ruff==0.14.11
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.41
tinycss2==1.4.0
tinyhtml5==2.0.0
twilio==9.6.5
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
vobject==0.9.6.1
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
xlrd==1.2.0
XlsxWriter==3.0.2
xlwt==1.3.0
yarl==1.20.1
zeep==4.1.0
zope.event==5.1
zope.interface==7.2
zopfli==0.2.3.post1
